# Add GJson to benchmark.