--- a/Dockerfile
+++ b/Dockerfile
@@ -7,6 +7,7 @@ RUN go get github.com/antonholmquist/jason
 RUN go get github.com/mreiferson/go-ujson
 RUN go get -tags=unsafe -u github.com/ugorji/go/codec
 RUN go get github.com/mailru/easyjson
+RUN go get github.com/tidwall/gjson
 
 WORKDIR /go/src/github.com/buger/jsonparser
 ADD . /go/src/github.com/buger/jsonparser