@WilliamHeaven 2020-03-13
@ksimka 
Hello, how should I run the benchmark, at the moment I am running, I get an error. Do you know why?

# benchmarks [benchmarks.test]
.\benchmark_codecgen.go:70:6: r.EncodeArrayStart undefined (type codec.genHelperEncDriver has no field or method EncodeArrayStart)
.\benchmark_codecgen.go:78:6: r.EncodeMapStart undefined (type codec.genHelperEncDriver has no field or method EncodeMapStart)
.\benchmark_codecgen.go:82:6: z.EncSendContainerState undefined (type codec.genHelperEncoder has no field or method EncSendContainerState)
.\benchmark_codecgen.go:90:6: z.EncSendContainerState undefined (type codec.genHelperEncoder has no field or method EncSendContainerState)
.\benchmark_codecgen.go:92:6: z.EncSendContainerState undefined (type codec.genHelperEncoder has no field or method EncSendContainerState)
.\benchmark_codecgen.go:101:6: z.EncSendContainerState undefined (type codec.genHelperEncoder has no field or method EncSendContainerState)
.\benchmark_codecgen.go:109:6: z.EncSendContainerState undefined (type codec.genHelperEncoder has no field or method EncSendContainerState)
.\benchmark_codecgen.go:111:6: z.EncSendContainerState undefined (type codec.genHelperEncoder has no field or method EncSendContainerState)
.\benchmark_codecgen.go:120:6: z.EncSendContainerState undefined (type codec.genHelperEncoder has no field or method EncSendContainerState)
.\benchmark_codecgen.go:128:6: z.EncSendContainerState undefined (type codec.genHelperEncoder has no field or method EncSendContainerState)
.\benchmark_codecgen.go:128:6: too many errors